Refactor vendors & include 3pts to index.ts
