+++ jest.config.js
@@ -27,7 +27,7 @@ module.exports = {
 	// coverageDirectory: undefined,
 
 	// An array of regexp pattern strings used to skip coverage collection
-	coveragePathIgnorePatterns: ['src/third-party-tags/external-scripts'],
+	coveragePathIgnorePatterns: ['vendor'],
 
 	// Indicates which provider should be used to instrument code for coverage
 	// coverageProvider: "babel",
