+++ src/index.ts
@@ -1,3 +1,5 @@
 export { ias } from './third-party-tags/ias';
 export { permutive } from './third-party-tags/permutive';
 export { fbPixel } from './third-party-tags/facebook-pixel';
+export { twitter } from './third-party-tags/twitter-uwt';
+export { lotame } from './third-party-tags/lotame';
