+++ src/third-party-tags/twitter-uwt.ts
@@ -1,8 +1,8 @@
+import { twitterScript as insertSnippet } from '../../vendor/twitter-script';
 import { GetThirdPartyTag } from '../types';
-import { twitterScript as insertSnippet } from './external-scripts/twitter-script';
 
 export const twitter: GetThirdPartyTag = ({ shouldRun }) => ({
 	shouldRun,
 	name: 'twitter',
-	insertSnippet,
+	insertSnippet: insertSnippet,
 });
